Move some patches into `layer_state_patches.py`
